+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   Dashboard, in the FAT link list and in the FAT link details view as well. You can
   only close your own ESI fleets.
 
+### Changed
+
+- Manual FAT function limited. It is only available when the FAT link has been
+  created within the last 24 hours and has not been re-opened.
+
 
 ## [2.1.0] - 2021-04-22